refactor: extract form into scheduling form
